mediaplayerservice: reset mBytesWritten in AudioSink flush and stop
For non-offload cases, when repeat mode is set to repeat track,
post a seek operation, progress bar does not move properly after
playback of current track is complete.

When audioSink is re-opened to replay the track (some apps do not
set the looping flag due to which track is re-opened),
mAudioSink::getFramesWritten() returns an invalid value due to which
calculation of positionUs in getCurrentPosition goes awry.

Reset mBytesWritten in MediaPlayerService in flush() and stop()
so that the correct number of frames written is returned when
queried by NuPlayerRenderer.

Change-Id: I9596ea715379b64236408dce506d03f4de0b2cb1
CRs-Fixed: 866102